+++ perceval/components/__init__.py
@@ -39,7 +39,7 @@ from ._pauli import (PauliType, PauliEigenStateType, get_pauli_eigen_state_prep_
 from .tomography_exp_configurer import processor_circuit_configurator
 from .comp_utils import decompose_perms
 from .port import APort, Port, Herald, PortLocation, get_basic_state_from_ports
-from .detector import IDetector, DetectionType, Detector, BSLayeredPPNR, get_detection_type
+from .detector import IDetector, DetectionType, Detector, BSLayeredPPNR, get_detection_type, check_heralds_detectors
 from .unitary_components import BSConvention, BS, PS, WP, HWP, QWP, PR, Unitary, PERM, PBS, Barrier
 from .non_unitary_components import TD, LC
 from .component_catalog import Catalog
