+++ perceval/components/detector.py
@@ -37,6 +37,7 @@ from .abstract_component import AComponent
 from .linear_circuit import Circuit
 from .unitary_components import BS, PERM
 from perceval.utils import BasicState, BSDistribution
+from perceval.utils.logging import channel, get_logger
 
 
 class DetectionType(Enum):
@@ -55,7 +56,6 @@ DetectionType.Mixed.__doc__ = "Multiple DetectionType"
 
 
 class IDetector(AComponent, ABC):
-    max_value: int
 
     def __init__(self):
         super().__init__(1)
@@ -79,6 +79,11 @@ class IDetector(AComponent, ABC):
     def copy(self, subs=None) -> IDetector:
         return copy.copy(self)
 
+    @property
+    @abstractmethod
+    def max_detections(self):
+        pass
+
 
 class BSLayeredPPNR(IDetector):
     """
@@ -101,7 +106,7 @@ class BSLayeredPPNR(IDetector):
         self._cache = {}  # This cache records simulations for a given photon count to speed up computations
 
     @property
-    def max_value(self) -> int:
+    def max_detections(self) -> int:
         """Maximum number of detected photons"""
         return 2 ** self._layers
 
@@ -192,7 +197,7 @@ class Detector(IDetector):
         self._cache = {}
 
     @property
-    def max_value(self) -> int:
+    def max_detections(self) -> int:
         """Maximum number of detected photons (None for infinity)"""
         return self._max
 
@@ -287,3 +292,20 @@ def get_detection_type(detectors: list[IDetector]) -> DetectionType:
         elif result != current:
             return DetectionType.Mixed
     return result
+
+
+def check_heralds_detectors(heralds: dict[int, int] | None, detectors: list[IDetector | None] | None) -> bool:
+    """
+    Check that heralds are compatible with the given detectors.
+     Returns True if the maximum value for all detectors is bigger than the expected herald value
+    """
+    if heralds and detectors:
+        for k, v in heralds.items():
+            detector = detectors[k]
+            if detector:
+                max_val = detector.max_detections
+                if max_val is not None and max_val < v:
+                    get_logger().warn(f"Incompatible heralds and detectors on mode {k}", channel.user)
+                    return False
+
+    return True
