--- perceval/simulators/_simulation_checker.py
@@ -1,54 +1,0 @@
-# MIT License
-#
-# Copyright (c) 2022 Quandela
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# As a special exception, the copyright holders of exqalibur library give you
-# permission to combine exqalibur with code included in the standard release of
-# Perceval under the MIT license (or modified versions of such code). You may
-# copy and distribute such a combined system following the terms of the MIT
-# license for both exqalibur and Perceval. This exception for the usage of
-# exqalibur is limited to the python bindings used by Perceval.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-from __future__ import annotations
-
-from perceval.components import IDetector
-from perceval.utils.logging import channel, get_logger
-
-
-class SimulatorChecker:
-    """This class is a helper to simulator to check consistency on the simulation parameters"""
-
-    @staticmethod
-    def check_heralds_detectors(heralds: dict[int, int] | None, detectors: list[IDetector | None] | None) -> bool:
-        """
-        Check that heralds are compatible with the given detectors.
-         Returns True if the maximum value for all detectors is bigger than the expected herald value
-        """
-        if heralds and detectors:
-            for k, v in heralds.items():
-                detector = detectors[k]
-                if detector:
-                    max_val = detector.max_value
-                    if max_val is not None and max_val < v:
-                        get_logger().warn(f"Incompatible heralds and detectors on mode {k}", channel.user)
-                        return False
-
-        return True
