+++ tests/test_detectors.py
@@ -26,11 +26,16 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
+from unittest.mock import patch
+
 import pytest
 
 from perceval.components import BS, PERM
-from perceval.components.detector import get_detection_type, Detector, DetectionType, BSLayeredPPNR
+from perceval.components.detector import get_detection_type, Detector, DetectionType, BSLayeredPPNR, \
+    check_heralds_detectors
 from perceval.utils import BasicState
+from _test_utils import LogChecker
+import perceval as pcvl
 
 
 def test_detector():
@@ -149,3 +154,18 @@ def test_detection_type():
     # PPNR means mixed detectors in this context
     assert get_detection_type(pnr_detector_list + thr_detector_list) == DetectionType.Mixed
     assert get_detection_type(mixed_detector_list) == DetectionType.Mixed
+
+
+@patch.object(pcvl.utils.logging.ExqaliburLogger, "warn")
+def test_incompatible_heralds(mock_warn):
+    detector_th = Detector.threshold()
+    detector_2 = Detector.ppnr(24, 2)
+    detector_pnr = Detector.pnr()
+    assert check_heralds_detectors({0: 14, 1: 1, 2: 2}, [None, detector_th, detector_2])
+    assert check_heralds_detectors({0: 14, 1: 1, 2: 2}, [detector_pnr, detector_th, detector_2])
+
+    with LogChecker(mock_warn):
+        assert not check_heralds_detectors({0: 14, 1: 2, 2: 0}, [None, detector_th, detector_2])
+
+    with LogChecker(mock_warn):
+        assert not check_heralds_detectors({0: 14, 1: 1, 2: 3}, [None, detector_th, detector_2])
