--- tests/test_sim_checker.py
@@ -1,49 +1,0 @@
-# MIT License
-#
-# Copyright (c) 2022 Quandela
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# As a special exception, the copyright holders of exqalibur library give you
-# permission to combine exqalibur with code included in the standard release of
-# Perceval under the MIT license (or modified versions of such code). You may
-# copy and distribute such a combined system following the terms of the MIT
-# license for both exqalibur and Perceval. This exception for the usage of
-# exqalibur is limited to the python bindings used by Perceval.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-from unittest.mock import patch
-
-import perceval as pcvl
-from perceval import Detector
-from perceval.simulators._simulation_checker import SimulatorChecker
-from _test_utils import LogChecker
-
-
-@patch.object(pcvl.utils.logging.ExqaliburLogger, "warn")
-def test_incompatible_heralds_slos(mock_warn):
-    checker = SimulatorChecker()
-    detector_th = Detector.threshold()
-    detector_2 = Detector.ppnr(24, 2)
-    assert checker.check_heralds_detectors({0: 14, 1: 1, 2: 2}, [None, detector_th, detector_2])
-
-    with LogChecker(mock_warn):
-        assert not checker.check_heralds_detectors({0: 14, 1: 2, 2: 0}, [None, detector_th, detector_2])
-
-    with LogChecker(mock_warn):
-        assert not checker.check_heralds_detectors({0: 14, 1: 1, 2: 3}, [None, detector_th, detector_2])
